Extract with_retries into a module
As per conversation between DP and PDB in
https://github.com/ministryofjustice/prison-visits-2/pull/39.  From DP:
"Each step is being forced to have behaviour for the sake of a consistent
interface when not all of the steps are the same".
