+++ smoke_test/smoke_test.rb
@@ -3,6 +3,7 @@ require 'capybara/dsl'
 require 'phantomjs/poltergeist'
 require 'capybara/poltergeist'
 
+require_relative 'helpers/with_retries'
 require_relative 'state'
 require_relative 'mail_box'
 require_relative 'steps/base_step'
