+++ smoke_test/steps/process_visit_request_page.rb
@@ -1,6 +1,7 @@
 module SmokeTest
   module Steps
     class ProcessVisitRequestPage < BaseStep
+
       PAGE_PATH = %r{/prison/visits/([0-9a-f-]){36}}
 
       def validate!
